+++ src/pages/login/ui/styles.css
@@ -1,7 +1,7 @@
 .login-page {
   display: flex;
   justify-content: center;
-  padding: 60px 0;
+  padding: 100px 0;
   background-color: rgb(0, 0, 0);
 }
