+++ src/widgets/header/index.ts
@@ -1,1 +1,3 @@
 export { WideScreenHeader } from "./wide-screen-header";
+export { BurgerMenu } from "./burger-menu";
+export { Header } from "./header";
